Linux compatibility: Guard Windows-only modules with cfg macro
Allow building and using com-rs - in particular the macros and interface
helpers - on platforms like Linux, to call into (native) libraries
which expose their functionality through COM. Here it is not necessary
to register or retrieve COM classes on the system as these merely define
virtual function layout of instances in the same process.

This brings more feature parity with the deprecated com_rs [1] crate,
which already provided said functionality.

[1]: https://crates.io/crates/com-rs